requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.12=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
attrs=25.1.0=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2025.1.31=h56e8100_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.4.1=pypi_0
colorama=0.4.6=pypi_0
distro=1.9.0=pypi_0
frozenlist=1.5.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
jiter=0.8.2=pypi_0
libexpat=2.6.4=he0c23c2_0
libffi=3.4.2=h8ffe710_5
liblzma=5.6.4=h2466b09_0
libsqlite=3.48.0=h67fdade_1
libzlib=1.3.1=h2466b09_2
loguru=0.7.3=pypi_0
multidict=6.1.0=pypi_0
numpy=2.2.2=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.28.0=pypi_0
openssl=3.4.0=ha4e3fda_1
outcome=1.3.0.post0=pypi_0
pandas=2.2.3=pypi_0
pip=25.0=pyh8b19718_0
propcache=0.2.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pyexecjs=1.5.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.11.11=h3f84c4b_1_cpython
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.1=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
schedule=1.2.2=pypi_0
selenium=4.28.1=pypi_0
setuptools=75.8.0=pyhff2d567_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
tk=8.6.13=h5226925_1
tqdm=4.67.1=pypi_0
trio=0.28.0=pypi_0
trio-websocket=0.11.1=pypi_0
tweepy=4.15.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
ucrt=10.0.22621.0=h57928b3_1
urllib3=2.3.0=pypi_0
vc=14.3=h5fd82a7_24
vc14_runtime=14.42.34433=h6356254_24
vs2015_runtime=14.42.34433=hfef2bbc_24
websocket-client=1.8.0=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
win32-setctime=1.2.0=pypi_0
wsproto=1.2.0=pypi_0
yarl=1.18.3=pypi_0
